Fix transposed arguments to memset()
Arguments to memset calls were transposed, effectively setting zero bytes at
the given address to sizeof(m). Also, the "nitems" member of Property was an
int, although the type received from Xlib is unsigned long, which resulted in
GCC emitting a warning about a narrowing conversion - this is now fixed.